refactor: moved validation and steane star logical operators to bloqade.gemini.star
